move attrs to separate file
